Rename ValidExpression => Expression
